model_backends/multimodal_backend/main.py
"""多模态模型后端 - 示例实现

此示例展示如何实现一个多模态模型后端,处理RGB图像、深度图像和眼动数据。

环境要求:
- Python 3.9+
- PyTorch 1.13+
- torchvision 0.14+
- pillow 9.0+
- websockets 10.0+

安装依赖:
```bash
pip install -r requirements.txt
```

启动服务:
```bash
python main.py
```
"""

import base64
import io
import logging
import sys
from pathlib import Path
from typing import Any, Dict

# 添加base路径,以便导入BaseModelBackend
sys.path.insert(0, str(Path(__file__).parent.parent / "base"))

from base_backend import BaseModelBackend

# 按需导入模型依赖
try:
    import torch
    import torchvision.transforms as transforms
    from PIL import Image
    import numpy as np
except ImportError as e:
    print("错误: 缺少依赖库")
    print("请安装: pip install torch torchvision pillow numpy")
    sys.exit(1)


class MultimodalBackend(BaseModelBackend):
    """多模态模型后端实现
    
    功能:
    - 处理RGB图像、深度图像和眼动数据
    - 输出疲劳度分数、注意力水平、姿态状态等
    
    TODO: 替换为实际的多模态模型
    """
    
    async def initialize_model(self) -> None:
        """加载多模态模型"""
        self.logger.info("正在加载多模态模型...")
        
        # TODO: 加载实际的模型文件
        # 示例: self.model = torch.jit.load("models/multimodal_fatigue_v1.pt")
        # self.model.eval()
        
        # 图像预处理转换
        self.transform = transforms.Compose([
            transforms.Resize((224, 224)),
            transforms.ToTensor(),
            transforms.Normalize(
                mean=[0.485, 0.456, 0.406],
                std=[0.229, 0.224, 0.225]
            )
        ])
        
        # 模拟: 创建一个假的模型 (实际使用时删除此部分)
        self.model = "模拟模型 (请替换为实际模型)"
        
        self.logger.info("✅ 多模态模型加载完成")
        self.logger.warning("⚠️  当前使用模拟模型,请替换为实际训练的模型")
    
    async def process_inference(self, data: Dict[str, Any]) -> Dict[str, Any]:
        """执行多模态推理
        
        Args:
            data: 输入数据,包含:
                - rgb_frame: RGB图像 (base64编码的JPEG/PNG)
                - depth_frame: 深度图像 (base64编码,可选)
                - eye_data: 眼动数据 (可选)
                - metadata: 元数据 (时间戳等)
        
        Returns:
            推理结果:
                - fatigue_score: 疲劳度分数 (0-1)
                - attention_level: 注意力水平 (0-1)
                - pose_status: 姿态状态 ("good", "forward", "backward", "tilted")
                - blink_frequency: 眨眼频率
                - features: 特征向量 (可选,用于进一步分析)
        """
        rgb_b64 = data.get("rgb_frame")
        depth_b64 = data.get("depth_frame")
        metadata = data.get("metadata", {})
        
        if not rgb_b64:
            raise ValueError("缺少必需的RGB图像")
        
        # 1. 解码RGB图像
        try:
            rgb_bytes = base64.b64decode(rgb_b64)
            rgb_image = Image.open(io.BytesIO(rgb_bytes)).convert("RGB")
        except Exception as e:
            raise ValueError(f"RGB图像解码失败: {e}")
        
        # 2. 预处理RGB图像
        rgb_tensor = self.transform(rgb_image).unsqueeze(0)
        
        # 3. 处理深度图像 (如果有)
        depth_tensor = None
        if depth_b64:
            try:
                depth_bytes = base64.b64decode(depth_b64)
                depth_image = Image.open(io.BytesIO(depth_bytes)).convert("L")
                depth_array = np.array(depth_image) / 255.0
                depth_tensor = torch.from_numpy(depth_array).float().unsqueeze(0).unsqueeze(0)
            except Exception as e:
                self.logger.warning(f"深度图像解码失败,将忽略: {e}")
        
        # 4. 模型推理
        # TODO: 替换为实际的模型推理代码
        # with torch.no_grad():
        #     if depth_tensor is not None:
        #         output = self.model(rgb_tensor, depth_tensor)
        #     else:
        #         output = self.model(rgb_tensor)
        #     
        #     fatigue_score = output['fatigue'].item()
        #     attention_level = output['attention'].item()
        #     pose_logits = output['pose']
        #     features = output['features'].cpu().numpy().tolist()
        
        # 模拟推理结果 (实际使用时删除此部分)
        import random
        fatigue_score = random.uniform(0.3, 0.9)
        attention_level = random.uniform(0.4, 0.95)
        pose_status = random.choice(["good", "forward", "backward", "tilted"])
        blink_frequency = random.uniform(10, 25)
        
        # 5. 返回结果
        result = {
            "fatigue_score": round(fatigue_score, 3),
            "attention_level": round(attention_level, 3),
            "pose_status": pose_status,
            "blink_frequency": round(blink_frequency, 2),
            "has_depth": depth_tensor is not None,
            "image_size": rgb_image.size,
            # "features": features,  # 特征向量 (可选)
        }
        
        self.logger.debug(
            f"推理完成: 疲劳度={result['fatigue_score']}, "
            f"注意力={result['attention_level']}, 姿态={result['pose_status']}"
        )
        
        return result
    
    async def cleanup(self) -> None:
        """清理模型资源"""
        self.logger.info("正在清理多模态模型资源...")
        
        # 释放模型
        self.model = None
        
        # 清理GPU缓存 (如果使用GPU)
        if torch.cuda.is_available():
            torch.cuda.empty_cache()
        
        # 强制垃圾回收
        import gc
        gc.collect()
        
        self.logger.info("✅ 资源清理完成")


def main():
    """主函数"""
    # 配置日志
    logging.basicConfig(
        level=logging.INFO,
        format='%(asctime)s [%(levelname)s] %(name)s: %(message)s',
        datefmt='%Y-%m-%d %H:%M:%S'
    )
    
    # 模型后端配置
    config = {
        "model_type": "multimodal",
        "host": "127.0.0.1",
        "port": 8766
    }
    
    # 创建并运行后端
    backend = MultimodalBackend(config)
    
    print("\n" + "=" * 70)
    print("多模态模型后端")
    print("=" * 70)
    print(f"监听地址: ws://{config['host']}:{config['port']}")
    print("按 Ctrl+C 停止服务")
    print("=" * 70 + "\n")
    
    backend.run()


if __name__ == "__main__":
    main()
